Add exceptions to parser. Add function parsing and converting to skr

+++ src/spark/common/StringRef.cpp
@@ -1,7 +1,7 @@
 #include "StringRef.h"
 
 std::string operator+(const std::string& s1, StringRef s2) {
-    int bufSz = std::min(32, s2.getLength() + 1);
+    size_t bufSz = std::min((size_t) 32, s2.getLength() + 1);
     char* buf = (char*) alloca(bufSz);
     int len = s2.copyTo(buf, bufSz);
     return s1 + std::string(buf, len);
